+++ src/TehPers.FishingOverhaul/Config/ContentPacks/FishTraitsPack.cs
@@ -11,10 +11,15 @@ namespace TehPers.FishingOverhaul.Config.ContentPacks
     [JsonDescribe]
     public record FishTraitsPack : JsonConfigRoot
     {
+        /// <inheritdoc cref="JsonConfigRoot.Schema" />
+        protected override string Schema =>
+            $"{JsonConfigRoot.jsonSchemaRootUrl}contentPacks/fishTraits.schema.json";
+
         /// <summary>
         /// The fish traits to add.
         /// </summary>
-        public ImmutableDictionary<NamespacedKey, FishTraits> Add { get; init; } = ImmutableDictionary<NamespacedKey, FishTraits>.Empty;
+        public ImmutableDictionary<NamespacedKey, FishTraits> Add { get; init; } =
+            ImmutableDictionary<NamespacedKey, FishTraits>.Empty;
 
         /// <summary>
         /// Merges all the traits into a single content object.
@@ -22,7 +27,7 @@ namespace TehPers.FishingOverhaul.Config.ContentPacks
         /// <param name="content">The content to merge into.</param>
         public FishingContent AddTo(FishingContent content)
         {
-            return content with { SetFishTraits = content.SetFishTraits.AddRange(this.Add) };
+            return content with {SetFishTraits = content.SetFishTraits.AddRange(this.Add)};
         }
     }
 }