+++ src/TehPers.FishingOverhaul/Config/ContentPacks/TreasurePack.cs
@@ -10,10 +10,15 @@ namespace TehPers.FishingOverhaul.Config.ContentPacks
     [JsonDescribe]
     public record TreasurePack : JsonConfigRoot
     {
+        /// <inheritdoc cref="JsonConfigRoot.Schema" />
+        protected override string Schema =>
+            $"{JsonConfigRoot.jsonSchemaRootUrl}contentPacks/treasure.schema.json";
+
         /// <summary>
         /// The treasure entries to add.
         /// </summary>
-        public ImmutableArray<TreasureEntry> Add { get; init; } = ImmutableArray<TreasureEntry>.Empty;
+        public ImmutableArray<TreasureEntry> Add { get; init; } =
+            ImmutableArray<TreasureEntry>.Empty;
 
         /// <summary>
         /// Merges all the treasure entries into a single content object.
@@ -21,7 +26,7 @@ namespace TehPers.FishingOverhaul.Config.ContentPacks
         /// <param name="content">The content to merge into.</param>
         public FishingContent AddTo(FishingContent content)
         {
-            return content with { AddTreasure = content.AddTreasure.AddRange(this.Add) };
+            return content with {AddTreasure = content.AddTreasure.AddRange(this.Add)};
         }
     }
 }