+++ docs/TehPers.FishingOverhaul/CHANGELOG.md
@@ -22,6 +22,7 @@ Changelog for [Teh's Fishing Overhaul].
   `MaxChanceDepth` at the same time rather than only `MaxChanceDepth`.
 - Some records in the fishing API now have private constructors to ensure subtypes of them cannot be
   created aside from the ones built into the API.
+- Updated schema URL in the content pack docs to point to the raw schema.
 
 ### Fixed